src/requirements.txt
# Python dependencies for OpenClaw Host Management
typer==0.15.2
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
httpx==0.27.0
PyYAML==6.0.1
